+++ targets/TARGET_RENESAS/TARGET_RZ_A2XX/common/rza_io_regrw.h
@@ -84,7 +84,7 @@ extern void     RZA_IO_RegWrite_16(volatile uint16_t * ioreg, uint16_t write_val
  *                       If no one need to keep previous value or shift, use
  *                       IOREG_NONMASK_ACCESS instead.
  *****************************************************************************/
-extern void     RZA_IO_RegWrite_32(volatile uint32_t * ioreg, uint32_t write_value, uint32_t shift, uint32_t mask);
+extern void     RZA_IO_RegWrite_32(volatile unsigned long * ioreg, uint32_t write_value, uint32_t shift, uint32_t mask);
 
 /******************************************************************************
  * @fn            RZA_IO_RegRead_8
@@ -123,7 +123,7 @@ extern uint16_t RZA_IO_RegRead_16 (volatile uint16_t * ioreg, uint16_t shift, ui
  *                       IOREG_NONMASK_ACCESS instead.
  * @retval        read value from register
  *****************************************************************************/
-extern uint32_t RZA_IO_RegRead_32 (volatile uint32_t * ioreg, uint32_t shift, uint32_t mask);
+extern uint32_t RZA_IO_RegRead_32 (volatile unsigned long * ioreg, uint32_t shift, uint32_t mask);
 
 #ifdef __cplusplus
 }
